+++ repomatic/__init__.py
@@ -17,5 +17,5 @@
 
 from __future__ import annotations
 
-__version__ = "6.12.0"
+__version__ = "6.12.1.dev0"
 __git_tag_sha__ = ""